lesson20/requirements.txt
psycopg2-binary
python-dotenv
streamlit
requests
schedule
streamlit-autorefresh